Roslyn side changes for the new toggle button added to the Error list… (#11439)
* Roslyn side changes for the new toggle button added to the Error list to turn on/off full solution analysis

We are moving the "Enable full solution analysis" option from Tools Option to the error list for improved discoverability, while also making it language agnostic. This change ensures that we set the Roslyn full solution analysis options when this button is toggled in the error list. Additionally, when we identify that the error list toggle button is available (Dev14 U3 or later), we will hide the tools option checkbox for "Enable full solution analysis" and show a note about it being moved to the error list. Note that we still continue with language specific defaults (C# off, VB on), but once the user has hit the error list toggle, the option setting is consistent across all languages.

Addresses workitem https://github.com/dotnet/roslyn/issues/11392

* Address PR feedback from Heejae - localization of strings.

* Address PR feedback from Heejae

* Address PR feedback

* Move to newer MS.VS.Shell.14.0 dll to consume the new error list changes.

* Move the code accessing the new IErrorList2 type into a separate method marked with [MethodImpl(MethodImplOptions.NoInlining)]