Update RoslynSemanticVersion to 1.2.3 in the microupdate branch.
The actual semantic version for the last micro update that we released from this branch for Update 2 was 1.2.2. However, we haven't been updating the version with each released micro update - RoslynSemanticVersion was still set to 1.2.0. And this meant we were still generating nuget packages with semantic version 1.2.0 in our builds - we would then manually fix up the version before releasing these nugets.

Update the version to 1.2.3 so that the version will be correct in case we end up having to spin any further micro updates from this branch for Update 2.

We should do the same for Update 3 (i.e. update semantic version to 1.3.1 as soon as Update 3 has shipped).