Add a new API which can be used to store metadata for persistence
One current challenge of our option persistence story is the
determination of the name of an option being serialized is mixed in
with the core logic for how to serialize things to a store. This results
in some very hard-to-maintain code since the logic for an option
is spread across many different areas.