Reimplement our option serializers to consume the metadata on options
Now that our existing options have been tagged with the appropriate
metadata, we can use that metadata to drive serialization. This
updates IOptionSerializers to make them no longer have any language-
specific metadata, and are simply driven from the metadata included
on the option.

This also renames IOptionSerializer to IOptionPersister, which is
a better name.