Cut short recursion that can occur (by spec!) in better betternes
This handles the simplest cases that are likely to arise in practice. We simply limit
direct recursion in "better conversion target" to a depth of 100.
See #12061 for the related spec bug. The spec still must be adjusted to say what should happen here,
and the implementation will need to be changed to follow that new specification.