Produce temporary files for "no output" projects
The Visual Studio project system doesn't handle projects that don't
produce outputs very well. This problem was made a bit worse in
Update 3, and so it's easiest to just produce temporary outputs
that aren't consumed anywhere. Since we remove the setting of
TargetPath, we can't accidentally consume anything from these projects
in the real build.