Implements simple cases of tuple name roundtripping (#12344)
For the most part, this change implements the same encoding
for tuples that we use for dynamic. This means that interface
implementations and generic constraints are not covered, but
base types and type parameters are.

This change also doesn't have full tests for "long" tuples (nested tuple
transformations) or tuple types that can only be constructed in
metadata.