Bring back "UseHostObjectToExecute" to Vbc task...
This code path was never ported forward to the new version of the Vbc
task, because it wasn't needed for Roslyn scenarios (IDE HostObject
returns 'true' from IsDesignTime, so it short-circuits above, and real
builds always "UseAlternateToolToExecute").

For test impact builds, it would be useful to bring back the capability to
call Emit in the same process as MSBuild.