Wait 5 seconds before reloading metadata on a change
Fixes #12189, and restores the behavior of the native C# language service.

The problem here is that we frequently get the file change notification
while something is still writing the file, and so we get an IOException when
we try to read it, resulting in us caching empty metadata and reporting
spurious errors.