Make cibuild.sh --release work
The --release flag was being ignored to cibuild.sh because the define we
pass to Make was set before we actually checked the command line arguments
to cibuild.sh.  There was also a casing typo in a variable name the script
was using which caused the BUILD_CONFIGURATION setting to never be passed
when building the non bootstrap projects.