Activate the file in RenameFile scenario.
In Rename File to match type name scenario, the file rename is performed
by a remove followed by a add new. This causes the file to be closed and
reopened. But simply re-opening is not sufficient, we should make it be
the active document, because it was the active document before the
refactoring started. Without this, the file is opened in the background
and that switches away the user's context from the file they were
actively editing.