Code Model: Implement CodeNamespace.Comment and CodeNamespace.DocComment properties
This change properly implements the CodeNamespace.Comment and CodeNamespace.DocComment properties. Historically, C# would throw an exception if either of these properties were accessed on a CodeNamespace, but VB allowed them to work. With this change, the properties will work for *both* C# and VB (even though DocComment doesn't make much sense). Technically, this is a breaking change for C# in the sense that an exception will no longer be thrown. However, I expect that it will generally make life a bit easier for Code Model consumers by keeping the implementations consistent between C# and VB.