Enable the OutOfProc server by default.
This is necessary now that several IDE features (i.e. FindReferences/NavigateTo) use OutOfProc by default.
If the server is not on, these guys just fall back and compute everything in-proc, defeating the point.

This also addresses issues where OutOfProc wouldn't be used even when it was supposed to.  For example,
if 'CodeLens' was off, but was then turned on, then OutOfProc still wouldn't be used.