Each kind of compiler should take its /version info from its containing assembly's FileVersion
It appears we were lucky things worked before (that is, before I broke things by making the
output use file version and the tests use assembly version), as the implementation and tests
for /help disagreed on which assembly to use to get the version info. They were usually the
same information. We now consistently use the file version for the file containing the assembly
of the compiler main class.
Fixes #14172