Validate our bootstrap phase loads the correct binaries
Our bootstrap phase relies on loading Microsoft.Build.Tasks.CodeAnalysis from a specific place on disk.  There is no way in MSBuild to guarantee that suceeds.  It's possible for instance for MSBuild to have already loaded a DLL of the same name from a new location.  Or even if our DLL is loaded, it's possible that some of the downstream depenedencies fell prey to the same problem.

This adds a validation layer to our bootstrap phase to guard against that.  It verifies that our build task and dependencies are loading from the location that we expect them to load from.

Note: there is actually a bug in MSBuild right now that prevents us from correctly doing this in MSBuild 15.0.  This check will help us validate that bug was fixed properly. https://github.com/Microsoft/msbuild/issues/1183