Make fetching of logging stuff cheaper
We were previously realizing the workspace and then reading in all the
FunctionIds to see if we should log each of those. We can do that from
the global option service directory (avoiding having to touch the
Workspace) and also not doing that until we actually need to. In most
cases ETW is off so we shouldn't even be looking at all.