Don't crash on an empty target path in Lightweight solution load
If the indexing service fails to determine the commmand line arguments (for example because the
targeting pack is not installed), it passes us an empty string for the TargetPath.  Our code wasn't
prepared to handle an empty string, so convert it to null instead, which we are prepared to handle.

Fixes internal bugs 269571 and 274711.