Don't crash if we're given a project path that doesn't exist
In some cases (for example internal issue 273128, which is now fixed,
where the index would be return stale data if a solution was moved on disk),
we get project paths that aren't part of the solution and don't exist on disk.

Be resilient to this happening, even though we will end up trying to parse files
from their old locations, and won't recognize files from their actual locations.

Fixes internal issue 273134