Fix some inconsistencies in our project files
These changes don't matter for command line builds, but the project
system interprets empty property groups to define which configuration/
platform combinations exist. Thus, those should line up with the intent
of the project. If there was ambiguity, I've resolved in favor of the
project.json statement of the world since that's what was actually
happening.