Don't crash if you rename a file and change or remove the extension
If you use Solution Explorer to rename a file and the extension were
to change or get deleted, we might crash. That's really bad, and we
shouldn't do that.

The problem was another fun example where different systems might see
the rename first: if the document provider saw the RDT rename first,
it'd close a document that the MiscellaneousFilesWorkspace was tracking.
However, it saw things after the rename and in one place but hadn't
seen the rename yet from the RDT.