Ensure Visual Studio workspace events are raised on the UI thread
Previously, we'd raise workspace events on whatever synchronization
context created the workspace. If you created the VisualStudioWorkspace
on a background thread, then this would mean we'd raise events there
which is not expected by many parts of our codebase. This corrects that
ensuring they're always raised where they should be. This also applies
to all workspaces in Visual Studio, which has been a source of bugs
when somebody had another workspace that they created themselves
which broke expectations.

While doing this I simplifed the APIs to clarify what they were being
used for and also removed some FactoryFactories that were unnecessary.