Remove the old bad merge check
It seems it's possible for our bad branch commit to be missing.  Suspect
this is due to it being GC'd out of our history since it's no longer
used.

This check was to ensure we didn't accidentally merge back in a bad
change we forced pushed out of our repo.  Seems like it's no longer
needed.